+++ common/src/main/java/com/yanny/ali/plugin/common/nodes/LootPoolNode.java
@@ -1,7 +1,10 @@
 package com.yanny.ali.plugin.common.nodes;
 
 import com.yanny.ali.Utils;
-import com.yanny.ali.api.*;
+import com.yanny.ali.api.IClientUtils;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.api.ITooltipNode;
+import com.yanny.ali.api.ListNode;
 import com.yanny.ali.plugin.common.NodeUtils;
 import com.yanny.ali.plugin.server.EntryTooltipUtils;
 import net.minecraft.network.RegistryFriendlyByteBuf;
@@ -19,7 +22,7 @@ public class LootPoolNode extends ListNode {
 
     private final ITooltipNode tooltip;
 
-    public LootPoolNode(List<ILootModifier<?>> modifiers, IServerUtils utils, LootPool lootPool, float chance, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
+    public LootPoolNode(IServerUtils utils, LootPool lootPool, float chance, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
         List<LootItemFunction> allFunctions = Stream.concat(functions.stream(), lootPool.functions.stream()).toList();
         List<LootItemCondition> allConditions = Stream.concat(conditions.stream(), lootPool.conditions.stream()).toList();
         int sumWeight = NodeUtils.getTotalWeight(lootPool.entries);